fix(ConfigFile): Add static_cast for signed/unsigned comparison (C4018)
- Fixed 2 C4018 warnings in IniFile.cpp lines 1622 and 1648
- Added static_cast<int>() for DWORD dwLiftPinNumberOfLn comparisons
- Ensures type safety in loop conditions

+++ SRC/ConfigFile/IniFile.cpp
@@ -1619,7 +1619,7 @@ void SaveNanoSpecIni(int iIniFile)
 				WritePrivateProfileString(section,szBuf,buff,szIniFilePath);
 			}
 #else
-			for (nIniStressPin=0; nIniStressPin<NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn; nIniStressPin++) {
+			for (nIniStressPin=0; nIniStressPin<static_cast<int>(NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn); nIniStressPin++) {
 				sprintf(szBuf,"LiftPinLocL%dPin%d_X",nIniStressLine+1,nIniStressPin+1);
 				wsprintf(buff,"%ld", NanoSysData.StressConfig.Line[nIniStressLine].PinPos[nIniStressPin].lX);
 				WritePrivateProfileString(section,szBuf,buff,szIniFilePath);
@@ -1645,7 +1645,7 @@ void SaveNanoSpecIni(int iIniFile)
 			if (NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn < 1)
 				continue;
 
-			for(nIniStressSection = 0;nIniStressSection < NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn - 1;nIniStressSection++){
+			for(nIniStressSection = 0;nIniStressSection < static_cast<int>(NanoSysData.StressConfig.Line[nIniStressLine].dwLiftPinNumberOfLn) - 1;nIniStressSection++){
 // 2009.09.17 bagus SE --}--
 				sprintf(szBuf,"L%dS%d_ScanStart_X",nIniStressLine+1,nIniStressSection+1);
 				wsprintf(buff,"%ld", NanoSysData.StressConfig.Line[nIniStressLine].SectPos[nIniStressSection].lScanStartPosX);
